# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Senne Platform Documentation'
  version: 1.0.0
paths:
...
  /api/patient/register:
    post:
      tags:
        - 'Register Patient'
      summary: 'Register Patient'
      description: 'Register Patient '
      operationId: 'Register Patient'
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
              properties:
                name:
                  type: text
                email:
                  type: email
                cpf:
                  type: text
                cnpj:
                  type: text
                phone:
                  type: text
              type: object
          application/json:
            schema: []
      responses:
        201:
          description: 'User registered successfully!'
          content:
            application/json:
              schema: []
        200:
          description: 'User registered successfully!'
          content:
            application/json:
              schema: []
        422:
          description: 'Unprocessable Entity'
          content:
            application/json:
              schema: []
        400:
          description: 'Bad request'
        401:
          description: Unauthorized
        404:
          description: 'Resource Not Found'
  /api/doctor/register:
    post:
      tags:
        - 'Register Doctor'
      summary: 'Register Doctor'
      description: 'Register Doctor '
      operationId: 'Register Doctor'
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
              properties:
                name:
                  type: text
                email:
                  type: email
                cpf:
                  type: text
                crm:
                  type: text
                phone:
                  type: text
                especialidade:
                  type: text
                novidades:
                  type: boolean
              type: object
          application/json:
            schema: []
      responses:
        201:
          description: 'User registered successfully!'
          content:
            application/json:
              schema: []
        200:
          description: 'User registered successfully!'
          content:
            application/json:
              schema: []
        422:
          description: 'Unprocessable Entity'
          content:
            application/json:
              schema: []
        400:
          description: 'Bad request'
        401:
          description: Unauthorized
        404:
          description: 'Resource Not Found'
  api/doctor/speciality:
    get:
      tags:
        - 'List Speciality '
      summary: Summary
      operationId: 4173b720efa5af10550da0610b260760
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
  'api/hospital/list/{id}':
    get:
      tags:
        - 'List Hospitals '
      summary: Summary
      operationId: 95f8e63945775467945587685ca0e386
      parameters:
        -
          name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
  api/hospital/list:
    get:
      tags:
        - 'List Hospitals '
      summary: Summary
      operationId: 901bdf8313532893b42d826467b54584
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
  /api/partner/register:
    post:
      tags:
        - 'Register Partner'
      summary: 'Register Partner'
      description: 'Register Partner '
      operationId: 'Register Partner'
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
              properties:
                name:
                  type: text
                email:
                  type: email
                cpf:
                  type: text
                phone:
                  type: text
                nameempresa:
                  type: text
                razaosocial:
                  type: text
                cnpj:
                  type: text
                classification:
                  type: text
                uf:
                  type: text
                cep:
                  type: text
                city:
                  type: text
                addres:
                  type: text
              type: object
          application/json:
            schema: []
      responses:
        201:
          description: 'solicitation sended'
          content:
            application/json:
              schema: []
        200:
          description: 'solicitation sended'
          content:
            application/json:
              schema: []
        422:
          description: 'Unprocessable Entity'
          content:
            application/json:
              schema: []
        400:
          description: 'Bad request'
        401:
          description: Unauthorized
        404:
          description: 'Resource Not Found'
  /api/user/hospital/register:
    post:
      tags:
        - 'Register User Hospital'
      summary: 'Register User Hospital'
      description: 'Register User Hospital '
      operationId: 'Register User Hospital'
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
              properties:
                name:
                  type: text
                email:
                  type: email
                cpf:
                  type: text
                phone:
                  type: text
                department:
                  type: text
                hospitals:
                  type: array
                  items: []
                permissions:
                  type: array
                  items: []
              type: object
          application/json:
            schema: []
      responses:
        201:
          description: 'User registered successfully!'
          content:
            application/json:
              schema: []
        200:
          description: 'User registered successfully!'
          content:
            application/json:
              schema: []
        422:
          description: 'Unprocessable Entity'
          content:
            application/json:
              schema: []
        400:
          description: 'Bad request'
        401:
          description: Unauthorized
        404:
          description: 'Resource Not Found'
  api/hospitals/domain/list:
    get:
      tags:
        - 'List Hospitals '
      summary: Summary
      operationId: 8acb1dde41998a6639e5c5c9bb494281
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - email
              properties:
                email:
                  type: email
              type: object
          application/json:
            schema: []
      responses:
        200:
          description: OK
        401:
          description: Unauthorized
        404:
          description: 'Not Found'
...